requirements.txt
qdrant-client>=1.9.0
openai>=1.30.0
langgraph>=0.2.0
langchain-core>=0.2.0
chainlit>=1.1.0
networkx>=3.3
python-dotenv>=1.0.0
rank-bm25>=0.2.2
pydantic>=2.7.0
tiktoken>=0.7.0
tqdm>=4.66.0
